requirements.txt
Flask
mysql-connector-python
bcrypt
bleach
opencv-python
numpy
ultralytics
gunicorn
python-dotenv